requirements-backend.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
websockets>=12.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
httpx>=0.27.0
gunicorn>=22.0.0
slowapi>=0.1.9
sentry-sdk[fastapi]>=2.0.0
bcrypt>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0
